+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.8.0] - Unreleased
+
+### Changed
+- Refactored the CLI. Run `rtlola-interpreter help` for more info.
+- Running the monitor in offline mode now requires the specification of the time format in the input source
 
 ## [0.7.0] - 23.11.2021